rename presidio_evaluator_2 to evaluator_2
